Make views class based, split them on separate files and put in new views folder.
